requirements.txt
pyTelegramBotAPI==3.6.6
python-telegram-bot==13.0